Uploaded new version
Updated the program to add several QoL improvements and bugfixes.   Fixed a major bug where the program would lag and hang as the key queue built up - it now only keeps the last 2 pressed keys in the queue. Added a new feature to load the default firefox profile instead of creating a new temporary profile - this allows you to use saved credentials and cookies, which must be set in a normal firefox session (not in the selenium instance spawned by rwkmulti)- thanks to Cagedangel for the suggestion.  This file uses a new naming convention (-latest) and will be used for the forthcoming auto-update feature.